app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
uvloop==0.19.0
orjson==3.9.15
supabase>=2,<3
python-jose[cryptography]==3.3.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx==0.24.1
pytest-mock>=3.11.0